+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8NQVQ0VFFGNBSS2FKAM4HM3"
+        "01K8Q55DAN3PKMG3940F8F2450"
     )
     r"""The unique identifier of the data source"""