+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-28T15:44:12.458Z")
+    created: Optional[datetime] = parse_datetime("2025-10-29T04:55:54.850Z")
     r"""The date and time the resource was created"""