+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8NQVQ0T3K2G88QSFFD9KAMH"
+        "01K8Q55DABS89XXJT7FM4AASV6"
     )
     r"""The unique identifier of the data source"""