+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8NQVQ0X2RP37VYWEJ0S8FFH"
+        "01K8Q55DAQSP7SZ0DPKFEVAKDA"
     )
     r"""The unique identifier of the data source"""